move invertMap to Lib
